Backend/services/base_service/requirements.txt
# Web app dependencies
fastapi==0.79.0
uvicorn==0.18.2
gunicorn==20.1.0
bidict==0.21.3
click==7.1.2
fastapi-socketio==0.0.9
httptools==0.3.0
pydantic==1.8.2
python-engineio==4.2.1
python-socketio==5.4.0
typing-extensions

# Data and calculation
pandas==1.4.1
numpy==1.22.2

# gcloud auth
grpcio==1.44.0
google-auth==2.6.0
google-cloud-secret-manager==2.8.0
google-cloud-storage==2.1.0
google-crc32c==1.3.0
google-resumable-media==2.2.1
google-api-core==2.5.0

# firestore
google-cloud-firestore==2.3.4

# External services
httpx==0.23.0
requests==2.26.0
requests-toolbelt==0.9.1

# Caching
cachetools==4.2.2

# Exponential backoff
backoff==1.11.1

# Logging
structlog==21.1.0

# Testing

protobuf==3.20.1
firebase-admin==5.2.0

python_multipart==0.0.5
phonenumbers==8.13.30


# send mail
sendgrid

# stripe
stripe

# mixpanel analytics
mixpanel

pydantic[email]